Fix path confusion on Windows (#1065)
This fixes an issue when running Slither on Windows. Some paths were
being stringified as Windows-style (backslash) while the rest of the
paths use POSIX-style (forward slash). This caused an issue when
trying to match paths, and Slither would complain the dependency
was not present.

After this change, the paths for import directives are always expressed
as POSIX-style paths when stringified.

See the following issue for a complete error message:

    https://github.com/crytic/crytic-compile/issues/241