fix: reorder named arguments to match declaration order (#1949)
Store names list in call operations and expressions. Reorder named arguments to match declaration order.

A function call with explicitly named parameters, of the form f({argName1: arg1, ...}), may list the parameters in an order different from that of the function's declaration. In this case, the IR should reorder the arguments to match the function declaration order before generating IR.

This PR implements named argument reordering for the following forms of function calls:
- struct constructors (most important)
- static internal calls

For the following forms of function calls, the order of the declaration arguments is not directly available during IR generation, so the PR did not implement reordering for them:
- external calls (HighLevelCall)
- library calls
- event calls

The get_declared_param_names function returns None for these unimplemented call forms.